Latent_Diffusion/requirements.txt
archisound==0.0.6
audio-diffusion-pytorch==0.1.3
auraloss==0.4.0
boto3==1.34.60
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
lightning==2.2.1
numpy==1.24.1
pandas==2.2.1
python-dotenv==1.0.1
pytorch-lightning==2.2.1
ruff==0.3.5
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
tqdm==4.66.2
wandb==0.16.4
